requirements.txt
click==8.1.7
Flask==2.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==2.1.1
Pillow==8.4.0
threadpoolctl==3.5.0
Werkzeug==2.0.3
